auto-claude: subtask-2-1 - Create useTheme hook for theme management
